+++ modular-particle-system/src/generators/circleEdgeGenerator.ts
@@ -20,6 +20,7 @@ import { loadSerializedProperty, deserializePrimitiveDataType } from "../seriali
  * center {
  *      @tooltip        TODO
  *      @type           Position
+ *      @defaultValue   { "x": 0, "y": 0 }
  * }
  * outerRadius {
  *      @tooltip        TODO
@@ -38,6 +39,7 @@ import { loadSerializedProperty, deserializePrimitiveDataType } from "../seriali
  * bursts {
  *      @tooltip        TODO
  *      @type           Burst[]
+ *      @defaultValue   []
  * }
  */
 export class CircleEdgeGenerator extends ParticleGenerator {
@@ -76,6 +78,7 @@ export class CircleEdgeGenerator extends ParticleGenerator {
         return {
             moduleTypeId: CircleEdgeGenerator.moduleTypeId,
             interval: this.interval,
+            bursts: this.bursts,
             center: this.center,
             outerRadius: this.outerRadius,
             innerRadius: this.innerRadius,