+++ modular-particle-system/src/generators/circleGenerator.ts
@@ -20,6 +20,7 @@ import { loadSerializedProperty, deserializePrimitiveDataType } from "../seriali
  * center {
  *      @tooltip        TODO
  *      @type           Position
+ *      @defaultValue   { "x": 0, "y": 0 }
  * }
  * radius {
  *      @tooltip        TODO
@@ -31,6 +32,7 @@ import { loadSerializedProperty, deserializePrimitiveDataType } from "../seriali
  * bursts {
  *      @tooltip        TODO
  *      @type           Burst[]
+ *      @defaultValue   []
  * }
  */
 export class CircleGenerator extends ParticleGenerator {
@@ -60,6 +62,7 @@ export class CircleGenerator extends ParticleGenerator {
         return {
             moduleTypeId: CircleGenerator.moduleTypeId,
             interval: this.interval,
+            bursts: this.bursts,
             center: this.center,
             radius: this.radius,
         };