+++ modular-particle-system/src/generators/generator.ts
@@ -13,8 +13,10 @@ export abstract class ParticleGenerator extends Module {
      *
      * `time` is seconds since the particle effect was created,
      * and `count` is number of particles to generate.
+     *
+     * Optional `repeat` property can be supplied to supply a second interval which the batch is automatically repeated with afterwards.
      */
-    bursts: Array<{ time: number; count: number }> = [];
+    bursts: Array<{ time: number; count: number; repeat?: number }> = [];
     private _updateCounter = 0;
 
     update(dt: number): void {
@@ -22,7 +24,12 @@ export abstract class ParticleGenerator extends Module {
         const tNow = this._updateCounter + dt;
         // Trigger all bursts that are between tPrev and tNow
         this.bursts.forEach((burst) => {
-            if (burst.time >= tPrev && burst.time <= tNow) {
+            const burstTimestamp =
+                burst.repeat === undefined || burst.repeat <= 0 || burst.time > tPrev
+                    ? burst.time
+                    : burst.time + burst.repeat * Math.ceil((tPrev - burst.time) / burst.repeat);
+
+            if (burstTimestamp >= tPrev && burstTimestamp <= tNow) {
                 for (let i = 0; i < burst.count; i += 1) {
                     this.generateParticle();
                 }