+++ modular-particle-system/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+-   Added optional `repeat` property to all generators `bursts` properties.
+    -   Can be used to make bursts repeat indefinitely.
+
 ## [1.0.0] - 2022-04-10
 
 ### Added